requirements.txt
flask
requests
aiohttp
python-dotenv
